pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "clamapp"
version = "0.1.0"
description = "A professional ClamAV GUI front-end for Linux desktops"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.10"
authors = [{ name = "ClamApp Contributors" }]
keywords = ["antivirus", "clamav", "security", "linux", "gui", "pyqt6"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "PyQt6>=6.4.0",
    "psutil>=5.9.0",
    "cryptography>=41.0.0",
    "Pillow>=10.0.0",
    "pyudev>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
]

[project.urls]
Homepage        = "https://github.com/YOUR_USERNAME/ClamApp"
"Bug Tracker"   = "https://github.com/YOUR_USERNAME/ClamApp/issues"
Releases        = "https://github.com/YOUR_USERNAME/ClamApp/releases"

# ── Entry point: `clamapp` command after `pip install .` ────────────────────
[project.scripts]
clamapp = "clamapp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["assets/icons/*", "*.desktop"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
